+++ src/main/java/com/example/QASystem/repositories/CheckListRepository.java
@@ -2,14 +2,13 @@ package com.example.QASystem.repositories;
 
 
 import com.example.QASystem.model.CheckList;
-import com.example.QASystem.model.TestCase;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CheckListRepository extends JpaRepository<CheckList, Integer> {
+public interface CheckListRepository extends JpaRepository<CheckList, Long> {
 
-    List<CheckList> getAllByProjectId(Integer id);
+    List<CheckList> getAllByProjectId(Long id);
 
 }
